Another rematch error fix

+++ CoreLogic.php
@@ -2134,9 +2134,9 @@ function NumChainLinks()
 
 function ClearGameFiles($gameName)
 {
-  unlink("./Games/" . $gameName . "/gamestateBackup.txt");
-  unlink("./Games/" . $gameName . "/beginTurnGamestate.txt");
-  unlink("./Games/" . $gameName . "/lastTurnGamestate.txt");
+  if(file_exists("./Games/" . $gameName . "/gamestateBackup.txt")) unlink("./Games/" . $gameName . "/gamestateBackup.txt");
+  if(file_exists("./Games/" . $gameName . "/beginTurnGamestate.txt")) unlink("./Games/" . $gameName . "/beginTurnGamestate.txt");
+  if(file_exists("./Games/" . $gameName . "/lastTurnGamestate.txt")) unlink("./Games/" . $gameName . "/lastTurnGamestate.txt");
 }
 
 function IsClassBonusActive($player, $class)
